fix(net-data): query ordering, JSON UTF-8 validity, CVE/country filters, Y2038, db robustness
Data-layer audit (storage format + query system), each verified on a real
4750-host dataset:

net-query:
- ORDER BY ip was lexicographic (ip is dotted-quad TEXT), scrambling results
  within each shard (305/310 distinct IPs misordered, e.g. 1.190 before 1.92).
  Collect each shard's rows and sort by numeric ip_to_u32 before emit, mirroring
  net_report.cc; drop the lexicographic SQL ORDER BY.
- json_escape passed bytes >=0x80 through raw, so a non-UTF-8 banner or a web
  title truncated mid-codepoint produced invalid-UTF-8 JSON that strict parsers
  reject. Validate UTF-8 and substitute U+FFFD (matches the nlohmann
  error_handler_t::replace the other two exporters already use).
- --nq-cve used an unanchored substring LIKE, so CVE-2024-1 matched
  CVE-2024-100/-1234 and CVE-description prose. Anchor on the quoted id.
- country filter wrapped the column in UPPER(), defeating idx_hosts_country
  (full scan, per EXPLAIN). Compare the bare (uppercase-stored) column instead.

net_db:
- PRAGMA busy_timeout=5000 so a concurrent reader/writer blocks instead of
  returning SQLITE_BUSY and silently dropping a discovered open; surface any
  open that still fails to write in net_db_insert_opens_bulk.
- net_db_count_unenriched counted port-rows while net_db_get_unenriched returns
  DISTINCT ip; use COUNT(DISTINCT ip) so the host count and reserve agree.
- net_db_search_fts only indexes rows with at least one searchable field, so the
  per-query transient FTS build skips bare discovery rows (correctness-preserving
  interim mitigation; a persistent index is the real scale fix).

formatting / Y2038:
- iso8601_utc / json_write_scaninfo / yuril_export_write_scaninfo took long
  (32-bit on Win64), truncating timestamps past 2038; widen to long long and
  drop the (long) casts at the call sites.
- web_recon extract_title hard-capped a title at 200 bytes mid-UTF-8; trim a
  trailing partial codepoint so the stored title stays valid UTF-8.

Verified: MSVC Win32 build clean; on real data numeric ordering (text + JSON),
valid-UTF-8 JSON export, --nq-cve no longer prefix-collides, country count
preserved via the index, FTS results unchanged.

Larger schema-migration items (ip_u32 column for indexed CIDR + ordering,
persistent FTS index, derived max_cvss/device_class/needs_enrich columns,
tls_not_after epoch normalization, entity-graph cohort truncation signal) are
documented as follow-ups, not in this commit.

+++ kmap.cc
@@ -2532,9 +2532,9 @@ int kmap_main(int argc, char *argv[]) {
   log_write(LOG_NORMAL | LOG_MACHINE, "%s %s scan initiated %s as: %s", KMAP_NAME, KMAP_VERSION, mytime, join_quoted(argv, argc).c_str());
   log_write(LOG_NORMAL | LOG_MACHINE, "\n");
   if (o.json_output_file)
-    json_write_scaninfo(KMAP_VERSION, join_quoted(argv, argc).c_str(), (long) timep);
+    json_write_scaninfo(KMAP_VERSION, join_quoted(argv, argc).c_str(), (long long) timep);
   if (o.yuril_export_dir)
-    yuril_export_write_scaninfo(KMAP_VERSION, join_quoted(argv, argc).c_str(), (long) timep);
+    yuril_export_write_scaninfo(KMAP_VERSION, join_quoted(argv, argc).c_str(), (long long) timep);
 
   /* Before we randomize the ports scanned, lets output them to machine
      parseable output */
